# splitter/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "pdf-dispatch",
    "version": "1.26.0",
    "description": "Self-hosted PDF splitting service. Watches a folder, splits multi-page\nPDFs on barcode/QR code detection, and routes output documents by trigger code.\n\nConfigurable entirely from the web interface; every setting is also accessible\nvia this REST API, making pdf-dispatch scriptable from Home Assistant, n8n,\nNode-RED, shell scripts, and any HTTP client.\n\n## Authentication\n\nAll endpoints except `/healthz` support two methods (checked in order):\n\n1. **API key** *(recommended for scripts)*: `X-API-Key: <key>` header.  \n   The key is auto-generated at startup and shown in the *API access* section\n   of the web interface. Override with the `API_KEY` environment variable.\n\n2. **HTTP Basic auth**: active only when both `APP_USERNAME` and\n   `APP_PASSWORD` environment variables are set.\n\nIf neither is configured the API is open (trusted network assumption).\nIf `X-API-Key` is present but wrong the request is **always rejected**\n— there is no fall-through to Basic auth.\n",
    "contact": {
      "name": "pdf-dispatch on GitHub",
      "url": "https://github.com/Lheriss/pdf-dispatch"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "/",
      "description": "This pdf-dispatch instance"
    }
  ],
...
  "security": [
    {
      "apiKey": []
    },
    {
      "basicAuth": []
    },
    {}
  ],
  "components": {
    "securitySchemes": {
      "apiKey": {
        "type": "apiKey",
        "in": "header",
        "name": "X-API-Key",
        "description": "Auto-generated 64-character hex key. Visible (and regeneratable) in the *API access* section of the web interface. Override with the `API_KEY` environment variable.\n"
      },
      "basicAuth": {
        "type": "http",
        "scheme": "basic",
        "description": "Active only when both `APP_USERNAME` and `APP_PASSWORD` environment variables are set.\n"
      }
    },
    "schemas": {
...
      "Trigger": {
        "type": "object",
        "required": [
          "value",
          "page_handling",
          "case_sensitive"
        ],
        "properties": {
          "value": {
            "type": "string",
            "description": "Exact code value or glob pattern (* ? [abc] [0-9]) to match against detected barcodes/QR codes.\n",
            "example": "INVOICE"
          },
          "page_handling": {
            "type": "string",
            "enum": [
              "keep",
              "delete"
            ],
            "description": "Combined with the global `separator_placement` option:\n- `keep + before` → trigger page is first page of output document\n- `keep + after`  → trigger page is last page of output document\n- `delete`        → trigger page removed regardless of placement\n",
            "default": "keep"
          },
          "case_sensitive": {
            "type": "boolean",
            "default": true
          }
        }
      }
    }
  },
  "paths": {
...
    "/api/state": {
      "get": {
        "summary": "Full application state",
        "description": "Main polling endpoint — queried every 3 s by the web interface.\nReturns in a single call: statistics, recent activity log, email\npolling status, and the complete sanitised application configuration\n(password hashes are never included).\n",
        "operationId": "getState",
        "tags": [
          "config"
        ],
        "responses": {
          "200": {
            "description": "Current state.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "stats": {
                      "type": "object",
                      "description": "Persistent processing statistics.",
                      "properties": {
                        "processed": {
                          "type": "integer"
                        },
                        "split_docs": {
                          "type": "integer"
                        },
                        "errors": {
                          "type": "integer"
                        },
                        "last_file": {
                          "type": "string",
                          "nullable": true
                        },
                        "last_time": {
                          "type": "string",
                          "nullable": true
                        }
                      }
                    },
                    "log": {
                      "type": "array",
                      "description": "Recent activity log (newest first).",
                      "items": {
                        "type": "object",
                        "properties": {
                          "level": {
                            "type": "string"
                          },
                          "message": {
                            "type": "string"
                          },
                          "time": {
                            "type": "string"
                          },
                          "file": {
                            "type": "string",
                            "nullable": true
                          }
                        }
                      }
                    },
                    "queue": {
                      "type": "array",
                      "description": "Filenames currently being processed.",
                      "items": {
                        "type": "string"
                      }
                    },
                    "app_config": {
                      "type": "object",
                      "description": "Complete application configuration. Key fields:\n",
                      "properties": {
                        "language": {
                          "type": "string",
                          "enum": [
                            "fr",
                            "en"
                          ]
                        },
                        "separator_placement": {
                          "type": "string",
                          "enum": [
                            "before",
                            "after"
                          ]
                        },
                        "split_values": {
                          "type": "array",
                          "items": {
                            "$ref": "#/components/schemas/Trigger"
                          }
                        },
                        "subdirs_by_trigger": {
                          "type": "boolean"
                        },
                        "delete_source": {
                          "type": "boolean"
                        },
                        "log_verbose": {
                          "type": "boolean"
                        },
                        "filename_separator": {
                          "type": "string"
                        },
                        "counter": {
                          "type": "integer"
                        },
                        "api_key": {
                          "type": "string"
                        },
                        "_api_key_env_set": {
                          "type": "boolean"
                        },
                        "webhook_enabled": {
                          "type": "boolean"
                        },
                        "webhook_url": {
                          "type": "string"
                        },
                        "webhook_events": {
                          "type": "string",
                          "enum": [
                            "all",
                            "success",
                            "error"
                          ]
                        }
                      }
                    },
                    "email": {
                      "type": "object",
                      "description": "Email polling summary."
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "Authentication required or invalid API key."
          }
        }
      }
    },
    "/api/config": {
      "post": {
        "summary": "Update configuration",
        "description": "Merges the provided object into the persistent configuration (top-level\nmerge only). For nested keys such as `dirs`, provide the complete\nupdated object.\n\nCommon use-cases:\n\n```json\n// Change language\n{\"language\": \"fr\"}\n\n// Update triggers\n{\"split_values\": [\n  {\"value\": \"INVOICE\", \"page_handling\": \"keep\", \"case_sensitive\": true}\n]}\n\n// Enable webhook\n{\"webhook_enabled\": true,\n \"webhook_url\": \"https://ha.local:8123/api/webhook/pdf\",\n \"webhook_events\": \"success\"}\n\n// Set separator placement\n{\"separator_placement\": \"after\"}\n```\n",
        "operationId": "updateConfig",
        "tags": [
          "config"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "description": "Partial configuration to merge."
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Complete updated configuration.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "config": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/config/validate_tokens": {
      "post": {
        "summary": "Validate filename token list",
        "description": "Validates a filename token configuration without persisting it.",
        "operationId": "validateTokens",
        "tags": [
          "config"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "tokens": {
                    "type": "array",
                    "items": {
                      "type": "object"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Validation result.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean"
                    },
                    "error": {
                      "type": "string",
                      "description": "Present when ok is false."
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/api/file/{filepath}": {
      "get": {
        "summary": "Download output file",
        "description": "Serves a file from OUTPUT_DIR. The `path` field returned by\nGET /api/recent and GET /api/tasks can be used directly as `{filepath}`.\n\nAccess is restricted to OUTPUT_DIR — the configuration file and\nin-progress input PDFs are never accessible.\n",
        "operationId": "downloadFile",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "name": "filepath",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            },
            "description": "Path relative to DATA_DIR (e.g. `output/INVOICE_001.pdf`).",
            "example": "output/INVOICE_001.pdf"
          },
          {
            "name": "download",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "0",
                "1"
              ],
              "default": "0"
            },
            "description": "`1` forces Content-Disposition: attachment (download dialog)."
          }
        ],
        "responses": {
          "200": {
            "description": "File content with appropriate Content-Type header.",
            "content": {
              "application/pdf": {
                "schema": {
                  "type": "string",
                  "format": "binary"
                }
              }
            }
          },
          "403": {
            "description": "Path resolves outside OUTPUT_DIR.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "404": {
            "description": "File not found.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/recent": {
      "get": {
        "summary": "List recent output files",
        "description": "Lists files in OUTPUT_DIR (excluding /error and /processed), sorted by modification time (newest first).\n",
        "operationId": "listRecentFiles",
        "tags": [
          "documents"
        ],
        "parameters": [
          {
            "name": "n",
            "in": "query",
            "schema": {
              "type": "integer",
              "default": 20,
              "maximum": 100
            },
            "description": "Maximum number of entries to return."
          }
        ],
        "responses": {
          "200": {
            "description": "File list.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "files": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/RecentFile"
                      }
                    },
                    "total": {
                      "type": "integer",
                      "description": "Total files found before truncation."
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Filesystem error.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/dirs/rename": {
      "post": {
        "summary": "Rename output folder",
        "description": "Renames a folder in the configuration and moves it on the filesystem.",
        "operationId": "renameDir",
        "tags": [
          "dirs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "key",
                  "path"
                ],
                "properties": {
                  "key": {
                    "type": "string",
                    "enum": [
                      "input",
                      "output",
                      "error",
                      "processed",
                      "no_code"
                    ]
                  },
                  "path": {
                    "type": "string",
                    "description": "New relative path (relative to DATA_DIR, no leading slash).",
                    "example": "output/invoices"
                  }
                }
              },
              "example": {
                "key": "no_code",
                "path": "output/unmatched"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Folder renamed successfully.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "config": {
                      "type": "object",
                      "description": "Updated dirs configuration."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Unknown key or invalid path.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Path already in use, destination exists, or processing in progress.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/api/dirs/recreate": {
      "post": {
        "summary": "Recreate deleted folder",
        "description": "Creates the folder on disk if it was accidentally deleted.",
        "operationId": "recreateDir",
        "tags": [
          "dirs"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "key"
                ],
                "properties": {
                  "key": {
                    "type": "string",
                    "enum": [
                      "input",
                      "output",
                      "error",
                      "processed",
                      "no_code"
                    ]
                  }
                }
              },
              "example": {
                "key": "no_code"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Folder recreated.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/OkTrue"
                }
              }
            }
          },
          "400": {
            "description": "Unknown folder key.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/settings/regenerate-api-key": {
      "post": {
        "summary": "Regenerate API key",
        "description": "Generates a new 64-character hex API key and persists it. The old\nkey stops working immediately.\n\nDisabled when the key is set via the `API_KEY` environment variable.\n",
        "operationId": "regenerateApiKey",
        "tags": [
          "system"
        ],
        "responses": {
          "200": {
            "description": "New key generated.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "ok": {
                      "type": "boolean",
                      "enum": [
                        true
                      ]
                    },
                    "key": {
                      "type": "string",
                      "description": "New API key (64-char hex)."
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Key is set via environment variable; regeneration disabled.",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  }
}